+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Each version should:
 - **Security**: to invite users to upgrade in case of vulnerabilities.
 
 ## [0.9.4] UNRELEASED
+### Changed
+- StructureExplorer.GetPathCheck() is like GetPath() but returns a second value, ok bool, to verify whether the value was found.
+- StructureExplorer.PathExists() returns true if a path points to a structure element with a non-nil value
+- Migrating some json checks to use the StructureExplorer interface instead of *gabs.Container. (Should probably make StructureExplorer its own subpackage.)
 
 ## [0.9.3] 2017-02-16
 ### Added